Jasho_backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
alembic
pytest
pytest-asyncio